ex1.m/trapezoidmaxerror.m: minors; report: ex.1.4 (cont.)

+++ src/ex1.m
@@ -58,30 +58,32 @@ x_1 = 5/2 * cos(0) + cos(9 * t + pi/2) + cos(13 * t + 3*pi/2) + 5/2 * cos(16 * t
 % Assuming t is a real variable, abs() is redundant; therefore, its removal
 % will provide the same result while being of faster computation, since it
 % difficults the integral calculation.
-% The lines commented below are also related with the abs function and its
-% integral.
+% The lines commented below are also related with warnings thrown by the
+% abs function and its 2nd order derivative (dirac function).
 
 % x_1_energy = abs(x_1) ^ 2;
 x_1_energy = x_1 ^ 2;
 
 % warning('off','symbolic:sym:int:warnmsg1'); % suppress the "Warning: Explicit integral could not be found." message
 tic;
-energy_t = double(int(x_1_energy, t, -pi, pi));
+energy_t = int(x_1_energy, t, -pi, pi);
 time_t = toc;
 % warning('on','symbolic:sym:int:warnmsg1');
-fprintf('x_1(t) exact energy value for t ∈ [-π,π]:\n\tvalue: %s ≈ %f\n\texecution_time: %fs\n\n', char(energy_t), double(energy_t), time_t);
+fprintf('x_1(t) exact energy value for t ∈ [-π,π]:\n\tvalue: %s ≈ %.4f\n\texecution_time: %fs\n\n', char(energy_t), double(energy_t), time_t);
 
 % warning('off','symbolic:mupadmex:MuPADTextWarning'); % suppress the "Warning: Function 'dirac' is not verified to be a valid MATLAB function." message
 tic;
 [energy_trap, step_trap] = trapezoidmaxerror(x_1_energy, -pi, pi, MAX_ERROR);
 time_trap = toc;
 % warning('on', 'symbolic:mupadmex:MuPADTextWarning');
-fprintf('x_1(t) approximated energy value for t ∈ [-π,π] using the Trapezoid Rule (maximal error of %.3f):\n\tvalue: %f\n\tnecessary step: %f\n\texecution time: %fs\n\n', MAX_ERROR, energy_trap, step_trap, time_trap);
+fprintf('x_1(t) approximated energy value for t ∈ [-π,π] using the Trapezoid Rule (maximal error of %.3f):\n\tvalue: %.4f\n\tnecessary step: %f\n\texecution time: %fs\n\n', MAX_ERROR, energy_trap, step_trap, time_trap);
 
+% warning('off','symbolic:mupadmex:MuPADTextWarning'); % suppress the "Warning: Function 'dirac' is not verified to be a valid MATLAB function." message
 tic;
 [energy_simpson, step_simpson] = simpsonmaxerror(x_1, -pi, pi, MAX_ERROR);
 time_simpson = toc;
-fprintf('x_1(t) approximated energy value for t ∈ [-π,π] using Simpson Rule (maximal error of %.3f):\n\tvalue: %f\n\tnecessary step: %f\n\texecution time: %fs\n', MAX_ERROR, energy_simpson, step_simpson, time_simpson);
+% warning('on', 'symbolic:mupadmex:MuPADTextWarning');
+fprintf('x_1(t) approximated energy value for t ∈ [-π,π] using Simpson Rule (maximal error of %.3f):\n\tvalue: %.4f\n\tnecessary step: %f\n\texecution time: %fs\n', MAX_ERROR, energy_simpson, step_simpson, time_simpson);
 
 %% Ex.1.5.
 %%%
